🤡 Add DIVA data quality validation integration mock

+++ tests/conftest.py
@@ -178,26 +178,30 @@ async def drop_all_tables() -> None:  # type: ignore[misc]
     try:
         yield
     finally:
-        # This logger is way too verbose
-        sa_logger = logging.getLogger("sqlalchemy.engine.Engine")
-        original_handlers = sa_logger.handlers
-        original_propagate = sa_logger.propagate
+        if await is_db_online_async():
+            # This logger is way too verbose
+            sa_logger = logging.getLogger("sqlalchemy.engine.Engine")
+            original_handlers = sa_logger.handlers
+            original_propagate = sa_logger.propagate
 
-        for handler in sa_logger.handlers[:]:
-            sa_logger.removeHandler(handler)
+            for handler in sa_logger.handlers[:]:
+                sa_logger.removeHandler(handler)
 
-        sa_logger.propagate = False
+            sa_logger.propagate = False
 
-        async with DBEngine.instance().begin() as conn:
-            metadata = sqlalchemy.MetaData()
-            await conn.run_sync(lambda sc: metadata.reflect(sc))
-            await conn.run_sync(lambda sc: metadata.drop_all(sc))
-            _logger.info("Dropped all tables")
+            try:
+                async with DBEngine.instance().begin() as conn:
+                    metadata = sqlalchemy.MetaData()
+                    await conn.run_sync(lambda sc: metadata.reflect(sc))
+                    await conn.run_sync(lambda sc: metadata.drop_all(sc))
+                    _logger.info("Dropped all tables")
+            except Exception:
+                _logger.warning("Failed to drop tables", exc_info=True)
 
-        for handler in original_handlers:
-            sa_logger.addHandler(handler)
+            for handler in original_handlers:
+                sa_logger.addHandler(handler)
 
-        sa_logger.propagate = original_propagate
+            sa_logger.propagate = original_propagate
 
 
 @pytest_asyncio.fixture(autouse=True, scope="function")
